# frontend/src/shared/api/generated/campaign-manager.openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RolePermissionsUpdate": {
        "properties": {
          "permissions": {
            "additionalProperties": {
              "type": "boolean"
            },
            "minProperties": 1,
            "title": "Permissions",
            "type": "object"
          }
        },
        "required": [
          "permissions"
        ],
        "title": "RolePermissionsUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Campaign CRUD, applicability filtering, A/B testing and analytics for the CashBack stack.",
    "title": "Cashback Campaign Manager",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/roles/permissions": {
      "get": {
        "operationId": "get_permissions_roles_permissions_get",
        "parameters": [
          {
            "in": "query",
            "name": "request",
            "required": true,
            "schema": {
              "title": "Request"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "additionalProperties": {
                      "type": "boolean"
                    },
                    "type": "object"
                  },
                  "title": "Response Get Permissions Roles Permissions Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Permissions",
        "tags": [
          "roles"
        ]
      }
    },
    "/roles/{role}/permissions": {
      "patch": {
        "operationId": "update_permissions_roles__role__permissions_patch",
        "parameters": [
          {
            "in": "path",
            "name": "role",
            "required": true,
            "schema": {
              "pattern": "^(ADMIN|MARKETER|ANALYST)$",
              "title": "Role",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "request",
            "required": true,
            "schema": {
              "title": "Request"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolePermissionsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Update Permissions Roles  Role  Permissions Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Permissions",
        "tags": [
          "roles"
        ]
      }
    }
  }
}